Shorten wording in order submenu tab headers
This commit shortens the wording in the tab headers so that the default
max container width will not force tabs to be in the dropdown.  We can
consider going back to the more verbose wording once the restrictive
grid system is changed. Adjust the tests for the new wording so that
the links are proper.