Prefer last level of breadcrumb over fallback page title
This change was mostly to move some logic out into a helper, but I also
prefer to see the last item in the breadcrumbs be the page title rather
than the controller name.

I would be fine with any page title pattern which isn't too long. We
might want to consider prepneding or appending something like
"Admin: " or "| Admin".